refactor!: remove search and filter functionality from sql queries (#48)
With the inclusion of Laravel Scout and the Meilisearch driver, searching and filtering resources via SQL queries is no longer necessary. The functionality and associated test have been removed from the system. This means that the `?search=` and `?{field}={?operator:}{value}` query parameters will no longer be present on any endpoints moving forward.

BREAKING CHANGE: Searching via SQL using query parameters has been deprecated. The `search` and `{field}` query parameters on resource endpoints will no longer work moving forward. Please use the `/v{apiVersion}/search` endpoint instead in order to search through the various resources.
