+++ tests/Feature/Endpoints/V0/ElementEndpointTest.php
@@ -87,86 +87,4 @@ class ElementEndpointTest extends TestCase
             ],
         ]);
     }
-
-    /** @test */
-    public function it_can_search_for_elements_via_the_name_column(): void
-    {
-        Element::factory()->create(['sort_id' => 1, 'name' => 'fire']);
-        $two = Element::factory()->create(['sort_id' => 3, 'name' => 'thunder'])->toArray();
-        $three = Element::factory()->create(['sort_id' => 7, 'name' => 'water'])->toArray();
-
-        $response = $this->get('/v0/elements?search=er');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'name' => $two['name'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'sort_id' => $three['sort_id'],
-                    'name' => $three['name'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_elements_via_the_name_column(): void
-    {
-        $one = Element::factory()->create(['sort_id' => 1, 'name' => 'fire'])->toArray();
-        Element::factory()->create(['sort_id' => 3, 'name' => 'thunder']);
-        Element::factory()->create(['sort_id' => 7, 'name' => 'water']);
-
-        $response = $this->get('/v0/elements?name=fire');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'name' => $one['name'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_elements_via_the_name_column_using_the_like_statement(): void
-    {
-        Element::factory()->create(['sort_id' => 1, 'name' => 'fire']);
-        $two = Element::factory()->create(['sort_id' => 3, 'name' => 'thunder'])->toArray();
-        $three = Element::factory()->create(['sort_id' => 7, 'name' => 'water'])->toArray();
-
-        $response = $this->get('/v0/elements?name=like:er');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'name' => $two['name'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'sort_id' => $three['sort_id'],
-                    'name' => $three['name'],
-                ],
-            ],
-        ]);
-    }
 }
