+++ tests/Feature/Endpoints/V0/SeedTestEndpointTest.php
@@ -89,83 +89,6 @@ class SeedTestEndpointTest extends TestCase
     }
 
     /** @test */
-    public function it_can_search_for_seed_tests_via_the_level_column(): void
-    {
-        $one = SeedTest::factory()->create(['level' => 1])->toArray();
-        SeedTest::factory()->create(['level' => 5]);
-        $three = SeedTest::factory()->create(['level' => 10])->toArray();
-
-        $response = $this->get('/v0/seed-tests?search=1');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'level' => $one['level'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'level' => $three['level'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_seed_tests_via_the_level_column(): void
-    {
-        $one = SeedTest::factory()->create(['level' => 1])->toArray();
-        SeedTest::factory()->create(['level' => 5]);
-        SeedTest::factory()->create(['level' => 10]);
-
-        $response = $this->get('/v0/seed-tests?level=1');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'level' => $one['level'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_seed_tests_via_the_level_column_using_the_like_statement(): void
-    {
-        $one = SeedTest::factory()->create(['level' => 1])->toArray();
-        SeedTest::factory()->create(['level' => 5]);
-        $three = SeedTest::factory()->create(['level' => 10])->toArray();
-
-        $response = $this->get('/v0/seed-tests?level=like:1');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'level' => $one['level'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'level' => $three['level'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
     public function it_can_load_the_test_questions_relation_on_a_list_of_seed_tests(): void
     {
         SeedTest::factory()
