+++ tests/Feature/Endpoints/V0/StatusEffectEndpointTest.php
@@ -91,397 +91,4 @@ class StatusEffectEndpointTest extends TestCase
             ],
         ]);
     }
-
-    /** @test */
-    public function it_can_search_for_status_effects_via_the_name_column(): void
-    {
-        StatusEffect::factory()->create([
-            'sort_id' => 1,
-            'name' => 'sleep',
-            'type' => 'harmful',
-        ]);
-        $two = StatusEffect::factory()->create([
-            'sort_id' => 2,
-            'name' => 'slow',
-            'type' => 'harmful',
-        ])->toArray();
-        StatusEffect::factory()->create([
-            'sort_id' => 3,
-            'name' => 'haste',
-            'type' => 'beneficial',
-        ]);
-
-        $response = $this->get('/v0/status-effects?search=slow');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'name' => $two['name'],
-                    'type' => $two['type'],
-                    'description' => $two['description'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_search_for_status_effects_via_the_type_column(): void
-    {
-        $one = StatusEffect::factory()->create([
-            'sort_id' => 1,
-            'name' => 'sleep',
-            'type' => 'harmful',
-        ])->toArray();
-        $two = StatusEffect::factory()->create([
-            'sort_id' => 2,
-            'name' => 'slow',
-            'type' => 'harmful',
-        ])->toArray();
-        StatusEffect::factory()->create([
-            'sort_id' => 3,
-            'name' => 'haste',
-            'type' => 'beneficial',
-        ]);
-
-        $response = $this->get('/v0/status-effects?search=harmful');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'name' => $one['name'],
-                    'type' => $one['type'],
-                    'description' => $one['description'],
-                ],
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'name' => $two['name'],
-                    'type' => $two['type'],
-                    'description' => $two['description'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_search_for_status_effects_via_the_description_column(): void
-    {
-        StatusEffect::factory()->create([
-            'sort_id' => 1,
-            'name' => 'sleep',
-            'type' => 'harmful',
-            'description' => 'Group One',
-        ]);
-        StatusEffect::factory()->create([
-            'sort_id' => 2,
-            'name' => 'slow',
-            'type' => 'harmful',
-            'description' => 'Group One',
-        ]);
-        $three = StatusEffect::factory()->create([
-            'sort_id' => 3,
-            'name' => 'haste',
-            'type' => 'beneficial',
-            'description' => 'Group Two',
-        ])->toArray();
-
-        $response = $this->get('/v0/status-effects?search=two');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $three['id'],
-                    'sort_id' => $three['sort_id'],
-                    'name' => $three['name'],
-                    'type' => $three['type'],
-                    'description' => $three['description'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_status_effects_via_the_name_column(): void
-    {
-        $one = StatusEffect::factory()->create([
-            'sort_id' => 1,
-            'name' => 'sleep',
-            'type' => 'harmful',
-        ])->toArray();
-        StatusEffect::factory()->create([
-            'sort_id' => 2,
-            'name' => 'slow',
-            'type' => 'harmful',
-        ]);
-        StatusEffect::factory()->create([
-            'sort_id' => 3,
-            'name' => 'haste',
-            'type' => 'beneficial',
-        ]);
-
-        $response = $this->get('/v0/status-effects?name=sleep');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'name' => $one['name'],
-                    'type' => $one['type'],
-                    'description' => $one['description'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_status_effects_via_the_name_column_using_the_like_statement(): void
-    {
-        $one = StatusEffect::factory()->create([
-            'sort_id' => 1,
-            'name' => 'sleep',
-            'type' => 'harmful',
-        ])->toArray();
-        $two = StatusEffect::factory()->create([
-            'sort_id' => 2,
-            'name' => 'slow',
-            'type' => 'harmful',
-        ])->toArray();
-        StatusEffect::factory()->create([
-            'sort_id' => 3,
-            'name' => 'haste',
-            'type' => 'beneficial',
-        ]);
-
-        $response = $this->get('/v0/status-effects?name=like:sl');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'name' => $one['name'],
-                    'type' => $one['type'],
-                    'description' => $one['description'],
-                ],
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'name' => $two['name'],
-                    'type' => $two['type'],
-                    'description' => $two['description'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_status_effects_via_the_type_column(): void
-    {
-        $one = StatusEffect::factory()->create([
-            'sort_id' => 1,
-            'name' => 'sleep',
-            'type' => 'harmful',
-        ])->toArray();
-        $two = StatusEffect::factory()->create([
-            'sort_id' => 2,
-            'name' => 'slow',
-            'type' => 'harmful',
-        ])->toArray();
-        StatusEffect::factory()->create([
-            'sort_id' => 3,
-            'name' => 'haste',
-            'type' => 'beneficial',
-        ]);
-
-        $response = $this->get('/v0/status-effects?type=harmful');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'name' => $one['name'],
-                    'type' => $one['type'],
-                    'description' => $one['description'],
-                ],
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'name' => $two['name'],
-                    'type' => $two['type'],
-                    'description' => $two['description'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_status_effects_via_the_type_column_using_the_like_statement(): void
-    {
-        $one = StatusEffect::factory()->create([
-            'sort_id' => 1,
-            'name' => 'sleep',
-            'type' => 'harmful',
-        ])->toArray();
-        $two = StatusEffect::factory()->create([
-            'sort_id' => 2,
-            'name' => 'slow',
-            'type' => 'harmful',
-        ])->toArray();
-        StatusEffect::factory()->create([
-            'sort_id' => 3,
-            'name' => 'haste',
-            'type' => 'beneficial',
-        ]);
-
-        $response = $this->get('/v0/status-effects?type=like:harm');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'name' => $one['name'],
-                    'type' => $one['type'],
-                    'description' => $one['description'],
-                ],
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'name' => $two['name'],
-                    'type' => $two['type'],
-                    'description' => $two['description'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_status_effects_via_the_description_column(): void
-    {
-        $one = StatusEffect::factory()->create([
-            'sort_id' => 1,
-            'name' => 'sleep',
-            'type' => 'harmful',
-            'description' => 'Group One',
-        ])->toArray();
-        $two = StatusEffect::factory()->create([
-            'sort_id' => 2,
-            'name' => 'slow',
-            'type' => 'harmful',
-            'description' => 'Group One',
-        ])->toArray();
-        StatusEffect::factory()->create([
-            'sort_id' => 3,
-            'name' => 'haste',
-            'type' => 'beneficial',
-            'description' => 'Group Two',
-        ]);
-
-        $response = $this->get('/v0/status-effects?description=Group%20One');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'name' => $one['name'],
-                    'type' => $one['type'],
-                    'description' => $one['description'],
-                ],
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'name' => $two['name'],
-                    'type' => $two['type'],
-                    'description' => $two['description'],
-                ],
-            ],
-        ]);
-    }
-
-    /** @test */
-    public function it_can_filter_status_effects_via_the_description_column_using_the_like_statement(): void
-    {
-        $one = StatusEffect::factory()->create([
-            'sort_id' => 1,
-            'name' => 'sleep',
-            'type' => 'harmful',
-            'description' => 'Group One',
-        ])->toArray();
-        $two = StatusEffect::factory()->create([
-            'sort_id' => 2,
-            'name' => 'slow',
-            'type' => 'harmful',
-            'description' => 'Group One',
-        ])->toArray();
-        StatusEffect::factory()->create([
-            'sort_id' => 3,
-            'name' => 'haste',
-            'type' => 'beneficial',
-            'description' => 'Group Two',
-        ]);
-
-        $response = $this->get('/v0/status-effects?description=like:one');
-
-        $response->assertStatus(200);
-        $response->assertExactJson([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'name' => $one['name'],
-                    'type' => $one['type'],
-                    'description' => $one['description'],
-                ],
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'name' => $two['name'],
-                    'type' => $two['type'],
-                    'description' => $two['description'],
-                ],
-            ],
-        ]);
-    }
 }
