+++ tests/Unit/Controllers/V0/SeedRankControllerTest.php
@@ -91,187 +91,4 @@ class SeedRankControllerTest extends TestCase
 
         $controller->show(new Request(), 'not-found');
     }
-
-    /** @test */
-    public function it_can_search_for_seed_ranks_via_the_rank_column(): void
-    {
-        $one = SeedRank::factory()->create(['rank' => '1', 'salary' => 500])->toArray();
-        SeedRank::factory()->create(['rank' => '5', 'salary' => 3000]);
-        $three = SeedRank::factory()->create(['rank' => '10', 'salary' => 8000])->toArray();
-
-        $service = $this->app->make(SeedRankService::class);
-        $transformer = new SeedRankTransformer();
-        $controller = new SeedRankController($service, $transformer);
-
-        $response = $controller->index(new Request(['search' => 1]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'rank' => $one['rank'],
-                    'salary' => $one['salary'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'rank' => $three['rank'],
-                    'salary' => $three['salary'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_search_for_seed_ranks_via_the_salary_column(): void
-    {
-        $one = SeedRank::factory()->create(['rank' => '1', 'salary' => 500])->toArray();
-        SeedRank::factory()->create(['rank' => '5', 'salary' => 3000]);
-        SeedRank::factory()->create(['rank' => '10', 'salary' => 8000]);
-
-        $service = $this->app->make(SeedRankService::class);
-        $transformer = new SeedRankTransformer();
-        $controller = new SeedRankController($service, $transformer);
-
-        $response = $controller->index(new Request(['search' => 500]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'rank' => $one['rank'],
-                    'salary' => $one['salary'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_seed_ranks_via_the_rank_column(): void
-    {
-        $one = SeedRank::factory()->create(['rank' => '1', 'salary' => 500])->toArray();
-        SeedRank::factory()->create(['rank' => '5', 'salary' => 3000]);
-        SeedRank::factory()->create(['rank' => '10', 'salary' => 8000]);
-
-        $service = $this->app->make(SeedRankService::class);
-        $transformer = new SeedRankTransformer();
-        $controller = new SeedRankController($service, $transformer);
-
-        $response = $controller->index(new Request(['rank' => 1]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'rank' => $one['rank'],
-                    'salary' => $one['salary'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_seed_ranks_via_the_rank_column_using_the_like_statement(): void
-    {
-        $one = SeedRank::factory()->create(['rank' => '1', 'salary' => 500])->toArray();
-        SeedRank::factory()->create(['rank' => '5', 'salary' => 3000]);
-        $three = SeedRank::factory()->create(['rank' => '10', 'salary' => 8000])->toArray();
-
-        $service = $this->app->make(SeedRankService::class);
-        $transformer = new SeedRankTransformer();
-        $controller = new SeedRankController($service, $transformer);
-
-        $response = $controller->index(new Request(['rank' => 'like:1']));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'rank' => $one['rank'],
-                    'salary' => $one['salary'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'rank' => $three['rank'],
-                    'salary' => $three['salary'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_seed_ranks_via_the_salary_column(): void
-    {
-        $one = SeedRank::factory()->create(['rank' => '1', 'salary' => 500])->toArray();
-        SeedRank::factory()->create(['rank' => '3', 'salary' => 1500]);
-        SeedRank::factory()->create(['rank' => '10', 'salary' => 8000]);
-
-        $service = $this->app->make(SeedRankService::class);
-        $transformer = new SeedRankTransformer();
-        $controller = new SeedRankController($service, $transformer);
-
-        $response = $controller->index(new Request(['salary' => 500]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'rank' => $one['rank'],
-                    'salary' => $one['salary'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_seed_ranks_via_the_salary_column_using_the_like_statement(): void
-    {
-        $one = SeedRank::factory()->create(['rank' => '1', 'salary' => 500])->toArray();
-        $two = SeedRank::factory()->create(['rank' => '3', 'salary' => 1500])->toArray();
-        SeedRank::factory()->create(['rank' => '10', 'salary' => 8000]);
-        $four = SeedRank::factory()->create(['rank' => '20', 'salary' => 15000])->toArray();
-
-        $service = $this->app->make(SeedRankService::class);
-        $transformer = new SeedRankTransformer();
-        $controller = new SeedRankController($service, $transformer);
-
-        $response = $controller->index(new Request(['salary' => 'like:500']));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'rank' => $one['rank'],
-                    'salary' => $one['salary'],
-                ],
-                [
-                    'id' => $two['id'],
-                    'rank' => $two['rank'],
-                    'salary' => $two['salary'],
-                ],
-                [
-                    'id' => $four['id'],
-                    'rank' => $four['rank'],
-                    'salary' => $four['salary'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
 }
