+++ tests/Unit/Controllers/V0/SeedTestControllerTest.php
@@ -93,92 +93,6 @@ class SeedTestControllerTest extends TestCase
     }
 
     /** @test */
-    public function it_can_search_for_seed_tests_via_the_level_column(): void
-    {
-        $one = SeedTest::factory()->create(['level' => 1])->toArray();
-        SeedTest::factory()->create(['level' => 5]);
-        $three = SeedTest::factory()->create(['level' => 10])->toArray();
-
-        $service = $this->app->make(SeedTestService::class);
-        $transformer = new SeedTestTransformer();
-        $controller = new SeedTestController($service, $transformer);
-
-        $response = $controller->index(new Request(['search' => 1]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'level' => $one['level'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'level' => $three['level'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_seed_tests_via_the_level_column(): void
-    {
-        $one = SeedTest::factory()->create(['level' => 1])->toArray();
-        SeedTest::factory()->create(['level' => 5]);
-        SeedTest::factory()->create(['level' => 10])->toArray();
-
-        $service = $this->app->make(SeedTestService::class);
-        $transformer = new SeedTestTransformer();
-        $controller = new SeedTestController($service, $transformer);
-
-        $response = $controller->index(new Request(['level' => 1]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'level' => $one['level'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_seed_tests_via_the_level_column_using_the_like_statement(): void
-    {
-        $one = SeedTest::factory()->create(['level' => 1])->toArray();
-        SeedTest::factory()->create(['level' => 5]);
-        $three = SeedTest::factory()->create(['level' => 10])->toArray();
-
-        $service = $this->app->make(SeedTestService::class);
-        $transformer = new SeedTestTransformer();
-        $controller = new SeedTestController($service, $transformer);
-
-        $response = $controller->index(new Request(['level' => 'like:1']));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'level' => $one['level'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'level' => $three['level'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
     public function it_can_load_the_test_questions_relation_on_a_list_of_seed_tests(): void
     {
         SeedTest::factory()
