+++ tests/Unit/Controllers/V0/TestQuestionControllerTest.php
@@ -75,458 +75,6 @@ class TestQuestionControllerTest extends TestCase
     }
 
     /** @test */
-    public function it_can_search_for_test_questions_via_the_question_number_column(): void
-    {
-        $one = TestQuestion::factory()->create([
-            'sort_id' => 1,
-            'question_number' => 1,
-            'question' => "Potions can restore a GF's HP.",
-            'answer' => false,
-        ])->toArray();
-        TestQuestion::factory()->create([
-            'sort_id' => 2,
-            'question_number' => 5,
-            'question' => 'Whoever strikes the finishing blow in battle receives the most EXP.',
-            'answer' => true,
-        ]);
-        $three = TestQuestion::factory()->create([
-            'question_number' => 10,
-            'question' => 'You can stock up to 255 of each magic.',
-            'answer' => false,
-        ])->toArray();
-
-        $service = $this->app->make(TestQuestionService::class);
-        $transformer = $this->app->make(TestQuestionTransformer::class);
-        $controller = new TestQuestionController($service, $transformer);
-
-        $response = $controller->index(new Request(['search' => 1]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'seed_test_id' => $one['seed_test_id'],
-                    'question_number' => $one['question_number'],
-                    'question' => $one['question'],
-                    'answer' => $one['answer'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'sort_id' => $three['sort_id'],
-                    'seed_test_id' => $three['seed_test_id'],
-                    'question_number' => $three['question_number'],
-                    'question' => $three['question'],
-                    'answer' => $three['answer'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_search_for_test_questions_via_the_question_column(): void
-    {
-        $one = TestQuestion::factory()->create([
-            'sort_id' => 1,
-            'question_number' => 1,
-            'question' => "Potions can restore a GF's HP.",
-            'answer' => false,
-        ])->toArray();
-        TestQuestion::factory()->create([
-            'sort_id' => 2,
-            'question_number' => 5,
-            'question' => 'Whoever strikes the finishing blow in battle receives the most EXP.',
-            'answer' => true,
-        ]);
-        $three = TestQuestion::factory()->create([
-            'sort_id' => 3,
-            'question_number' => 10,
-            'question' => 'You can stock up to 255 of each magic.',
-            'answer' => false,
-        ])->toArray();
-
-        $service = $this->app->make(TestQuestionService::class);
-        $transformer = $this->app->make(TestQuestionTransformer::class);
-        $controller = new TestQuestionController($service, $transformer);
-
-        $response = $controller->index(new Request(['search' => 'can']));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'seed_test_id' => $one['seed_test_id'],
-                    'question_number' => $one['question_number'],
-                    'question' => $one['question'],
-                    'answer' => $one['answer'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'sort_id' => $three['sort_id'],
-                    'seed_test_id' => $three['seed_test_id'],
-                    'question_number' => $three['question_number'],
-                    'question' => $three['question'],
-                    'answer' => $three['answer'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_search_for_test_questions_via_the_answer_column(): void
-    {
-        $one = TestQuestion::factory()->create([
-            'sort_id' => 1,
-            'question_number' => 1,
-            'question' => "Potions can restore a GF's HP.",
-            'answer' => false,
-        ])->toArray();
-        TestQuestion::factory()->create([
-            'sort_id' => 2,
-            'question_number' => 5,
-            'question' => 'Whoever strikes the finishing blow in battle receives the most EXP.',
-            'answer' => true,
-        ]);
-        $three = TestQuestion::factory()->create([
-            'sort_id' => 3,
-            'question_number' => 10,
-            'question' => 'You can stock up to 255 of each magic.',
-            'answer' => false,
-        ])->toArray();
-
-        $service = $this->app->make(TestQuestionService::class);
-        $transformer = $this->app->make(TestQuestionTransformer::class);
-        $controller = new TestQuestionController($service, $transformer);
-
-        $response = $controller->index(new Request(['search' => false]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'seed_test_id' => $one['seed_test_id'],
-                    'question_number' => $one['question_number'],
-                    'question' => $one['question'],
-                    'answer' => $one['answer'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'sort_id' => $three['sort_id'],
-                    'seed_test_id' => $three['seed_test_id'],
-                    'question_number' => $three['question_number'],
-                    'question' => $three['question'],
-                    'answer' => $three['answer'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_test_questions_via_the_question_number_column(): void
-    {
-        $one = TestQuestion::factory()->create([
-            'sort_id' => 1,
-            'question_number' => 1,
-            'question' => "Potions can restore a GF's HP.",
-            'answer' => false,
-        ])->toArray();
-        TestQuestion::factory()->create([
-            'sort_id' => 2,
-            'question_number' => 5,
-            'question' => 'Whoever strikes the finishing blow in battle receives the most EXP.',
-            'answer' => true,
-        ]);
-        TestQuestion::factory()->create([
-            'sort_id' => 3,
-            'question_number' => 10,
-            'question' => 'You can stock up to 255 of each magic.',
-            'answer' => false,
-        ]);
-
-        $service = $this->app->make(TestQuestionService::class);
-        $transformer = $this->app->make(TestQuestionTransformer::class);
-        $controller = new TestQuestionController($service, $transformer);
-
-        $response = $controller->index(new Request(['question_number' => 1]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'seed_test_id' => $one['seed_test_id'],
-                    'question_number' => $one['question_number'],
-                    'question' => $one['question'],
-                    'answer' => $one['answer'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_test_questions_via_the_question_number_column_using_the_like_statement(): void
-    {
-        $one = TestQuestion::factory()->create([
-            'sort_id' => 1,
-            'question_number' => 1,
-            'question' => "Potions can restore a GF's HP.",
-            'answer' => false,
-        ])->toArray();
-        TestQuestion::factory()->create([
-            'sort_id' => 2,
-            'question_number' => 5,
-            'question' => 'Whoever strikes the finishing blow in battle receives the most EXP.',
-            'answer' => true,
-        ]);
-        $three = TestQuestion::factory()->create([
-            'sort_id' => 3,
-            'question_number' => 10,
-            'question' => 'You can stock up to 255 of each magic.',
-            'answer' => false,
-        ])->toArray();
-
-        $service = $this->app->make(TestQuestionService::class);
-        $transformer = $this->app->make(TestQuestionTransformer::class);
-        $controller = new TestQuestionController($service, $transformer);
-
-        $response = $controller->index(new Request(['question_number' => 'like:1']));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'seed_test_id' => $one['seed_test_id'],
-                    'question_number' => $one['question_number'],
-                    'question' => $one['question'],
-                    'answer' => $one['answer'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'sort_id' => $three['sort_id'],
-                    'seed_test_id' => $three['seed_test_id'],
-                    'question_number' => $three['question_number'],
-                    'question' => $three['question'],
-                    'answer' => $three['answer'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_test_questions_via_the_question_column(): void
-    {
-        $one = TestQuestion::factory()->create([
-            'sort_id' => 1,
-            'question_number' => 1,
-            'question' => "Potions can restore a GF's HP.",
-            'answer' => false,
-        ])->toArray();
-        TestQuestion::factory()->create([
-            'sort_id' => 2,
-            'question_number' => 5,
-            'question' => 'Whoever strikes the finishing blow in battle receives the most EXP.',
-            'answer' => true,
-        ]);
-        TestQuestion::factory()->create([
-            'sort_id' => 3,
-            'question_number' => 10,
-            'question' => 'You can stock up to 255 of each magic.',
-            'answer' => false,
-        ]);
-
-        $service = $this->app->make(TestQuestionService::class);
-        $transformer = $this->app->make(TestQuestionTransformer::class);
-        $controller = new TestQuestionController($service, $transformer);
-
-        $response = $controller->index(new Request(['question' => "Potions can restore a GF's HP."]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'seed_test_id' => $one['seed_test_id'],
-                    'question_number' => $one['question_number'],
-                    'question' => $one['question'],
-                    'answer' => $one['answer'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_test_questions_via_the_question_column_using_the_like_statement(): void
-    {
-        $one = TestQuestion::factory()->create([
-            'sort_id' => 1,
-            'question_number' => 1,
-            'question' => "Potions can restore a GF's HP.",
-            'answer' => false,
-        ])->toArray();
-        TestQuestion::factory()->create([
-            'sort_id' => 2,
-            'question_number' => 5,
-            'question' => 'Whoever strikes the finishing blow in battle receives the most EXP.',
-            'answer' => true,
-        ]);
-        $three = TestQuestion::factory()->create([
-            'sort_id' => 3,
-            'question_number' => 10,
-            'question' => 'You can stock up to 255 of each magic.',
-            'answer' => false,
-        ])->toArray();
-
-        $service = $this->app->make(TestQuestionService::class);
-        $transformer = $this->app->make(TestQuestionTransformer::class);
-        $controller = new TestQuestionController($service, $transformer);
-
-        $response = $controller->index(new Request(['question' => 'like:can']));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'seed_test_id' => $one['seed_test_id'],
-                    'question_number' => $one['question_number'],
-                    'question' => $one['question'],
-                    'answer' => $one['answer'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'sort_id' => $three['sort_id'],
-                    'seed_test_id' => $three['seed_test_id'],
-                    'question_number' => $three['question_number'],
-                    'question' => $three['question'],
-                    'answer' => $three['answer'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_test_questions_via_the_answer_column(): void
-    {
-        TestQuestion::factory()->create([
-            'sort_id' => 1,
-            'question_number' => 1,
-            'question' => "Potions can restore a GF's HP.",
-            'answer' => false,
-        ]);
-        $two = TestQuestion::factory()->create([
-            'sort_id' => 2,
-            'question_number' => 5,
-            'question' => 'Whoever strikes the finishing blow in battle receives the most EXP.',
-            'answer' => true,
-        ])->toArray();
-        TestQuestion::factory()->create([
-            'sort_id' => 3,
-            'question_number' => 10,
-            'question' => 'You can stock up to 255 of each magic.',
-            'answer' => false,
-        ]);
-
-        $service = $this->app->make(TestQuestionService::class);
-        $transformer = $this->app->make(TestQuestionTransformer::class);
-        $controller = new TestQuestionController($service, $transformer);
-
-        $response = $controller->index(new Request(['answer' => true]));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $two['id'],
-                    'sort_id' => $two['sort_id'],
-                    'seed_test_id' => $two['seed_test_id'],
-                    'question_number' => $two['question_number'],
-                    'question' => $two['question'],
-                    'answer' => $two['answer'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
-    public function it_can_filter_test_questions_via_the_answer_column_using_the_like_statement(): void
-    {
-        $one = TestQuestion::factory()->create([
-            'sort_id' => 1,
-            'question_number' => 1,
-            'question' => "Potions can restore a GF's HP.",
-            'answer' => false,
-        ])->toArray();
-        TestQuestion::factory()->create([
-            'sort_id' => 2,
-            'question_number' => 5,
-            'question' => 'Whoever strikes the finishing blow in battle receives the most EXP.',
-            'answer' => true,
-        ]);
-        $three = TestQuestion::factory()->create([
-            'sort_id' => 3,
-            'question_number' => 10,
-            'question' => 'You can stock up to 255 of each magic.',
-            'answer' => false,
-        ])->toArray();
-
-        $service = $this->app->make(TestQuestionService::class);
-        $transformer = $this->app->make(TestQuestionTransformer::class);
-        $controller = new TestQuestionController($service, $transformer);
-
-        $response = $controller->index(new Request(['answer' => 'like:f']));
-
-        $this->assertEquals([
-            'success' => true,
-            'message' => 'Successfully retrieved data.',
-            'status_code' => 200,
-            'data' => [
-                [
-                    'id' => $one['id'],
-                    'sort_id' => $one['sort_id'],
-                    'seed_test_id' => $one['seed_test_id'],
-                    'question_number' => $one['question_number'],
-                    'question' => $one['question'],
-                    'answer' => $one['answer'],
-                ],
-                [
-                    'id' => $three['id'],
-                    'sort_id' => $three['sort_id'],
-                    'seed_test_id' => $three['seed_test_id'],
-                    'question_number' => $three['question_number'],
-                    'question' => $three['question'],
-                    'answer' => $three['answer'],
-                ],
-            ],
-        ], $response->getData(true));
-    }
-
-    /** @test */
     public function it_can_load_the_seed_test_relation_on_a_list_of_test_questions(): void
     {
         TestQuestion::factory()
