+++ tests/Unit/Services/SeedTestServiceTest.php
@@ -71,74 +71,6 @@ class SeedTestServiceTest extends TestCase
     }
 
     /** @test */
-    public function it_can_search_for_seed_tests_via_the_level_column(): void
-    {
-        $one = SeedTest::factory()->create(['level' => 1])->toArray();
-        SeedTest::factory()->create(['level' => 5]);
-        $three = SeedTest::factory()->create(['level' => 10])->toArray();
-
-        $model = new SeedTest();
-        $service = new SeedTestService($model);
-
-        $records = $service->all(new Request(['search' => 1]));
-
-        $this->assertEquals([
-            [
-                'id' => $one['id'],
-                'level' => $one['level'],
-            ],
-            [
-                'id' => $three['id'],
-                'level' => $three['level'],
-            ],
-        ], $records);
-    }
-
-    /** @test */
-    public function it_can_filter_seed_tests_via_the_level_column(): void
-    {
-        $one = SeedTest::factory()->create(['level' => 1])->toArray();
-        SeedTest::factory()->create(['level' => 5]);
-        SeedTest::factory()->create(['level' => 10]);
-
-        $model = new SeedTest();
-        $service = new SeedTestService($model);
-
-        $records = $service->all(new Request(['level' => 1]));
-
-        $this->assertEquals([
-            [
-                'id' => $one['id'],
-                'level' => $one['level'],
-            ],
-        ], $records);
-    }
-
-    /** @test */
-    public function it_can_filter_seed_tests_via_the_level_column_using_the_like_statement(): void
-    {
-        $one = SeedTest::factory()->create(['level' => 1])->toArray();
-        SeedTest::factory()->create(['level' => 5]);
-        $three = SeedTest::factory()->create(['level' => 10])->toArray();
-
-        $model = new SeedTest();
-        $service = new SeedTestService($model);
-
-        $records = $service->all(new Request(['level' => 'like:1']));
-
-        $this->assertEquals([
-            [
-                'id' => $one['id'],
-                'level' => $one['level'],
-            ],
-            [
-                'id' => $three['id'],
-                'level' => $three['level'],
-            ],
-        ], $records);
-    }
-
-    /** @test */
     public function it_can_load_the_test_questions_relation_on_a_list_of_seed_tests(): void
     {
         SeedTest::factory()
